Rename Nabvar component to Header component. Change App.css to App.scss
