+++ src/App.scss
@@ -1,0 +1,9 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+.header-right {
+  color: #0290ff;
+}
