--- src/components/Navbar.js
@@ -1,32 +1,0 @@
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-
-const Navbar = () => {
-  const links = [
-    {
-      id: 1,
-      path: '/',
-      text: 'Books',
-    },
-    {
-      id: 2,
-      path: '/categories',
-      text: 'Categories',
-    },
-  ];
-  return (
-    <nav className="navigation">
-      <h2 className="navbar-title">Bookstore ASD</h2>
-      <ul>
-        {links.map((link) => (
-          <li key={link.id}>
-            <NavLink to={link.path} activeClassName="active-link" exact>
-              {link.text}
-            </NavLink>
-          </li>
-        ))}
-      </ul>
-    </nav>
-  );
-};
-export default Navbar;
